refactor(product): change popularityScore from int to BigDecimal to support decimal values from JSON

+++ src/main/java/com/example/AssignmentBrief/dto/productDTO.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public record productDTO (
         String id,
         String name,
-        int popularityScore,
+        BigDecimal popularityScore,
         int weight,
         BigDecimal price,
         Map<String, String> images
